ID-Card-Detection/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

numpy==1.26.4
opencv-contrib-python==4.9.0.80

ultralytics==8.2.0

torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

python-multipart==0.0.9
python-docx==1.1.0
redis==5.0.8
